Extract canisterId resolution to shared util
This PR extracts the canisterId resolution mechanism used in the
VC flow into a shared utility. This utility will in an upcoming PR
also be used from the authentication flow.

This PR only _moves_ code.
